gnome-45: Port extension
