+++ .github/workflows/build.yml
@@ -14,13 +14,15 @@ jobs:
   build:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v4
+        with:
+          submodules: 'recursive'
       - uses: actions/setup-node@v2.4.0
       - name: Setup build dependencies
         run: |
           npm install
           sudo apt-get update
-          sudo apt-get install -y gnome-shell-extensions gettext
+          sudo apt-get install -y reuse gnome-shell-extensions gettext
 
       - name: Run linter
         run: make lint
