+++ src/metadata.json
@@ -5,6 +5,6 @@
   "description": "Replace the on-screen-display volume level icon with a number.",
   "uuid": "osd-volume-number@deminder",
   "url": "https://github.com/Deminder/osd-volume-number",
-  "shell-version": ["42", "43", "44"],
+  "shell-version": ["45"],
   "version": 6
 }
