+++ src/modules/util.js
@@ -1,0 +1,4 @@
+// SPDX-FileCopyrightText: 2023 Deminder <tremminder@gmail.com>
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+export { debugMode, logDebug } from './sdt/util.js';
